requirements.txt
annotated-types==0.7.0
attrs==24.2.0
beautifulsoup4==4.12.3
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
google-ai-generativelanguage==0.6.9
google-api-core==2.20.0
google-api-python-client==2.146.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.1
googleapis-common-protos==1.65.0
grpcio==1.66.1
grpcio-status==1.66.1
h11==0.14.0
httplib2==0.22.0
idna==3.10
outcome==1.3.0.post0
proto-plus==1.24.0
protobuf==5.28.2
pyaml==24.7.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
pyparsing==3.1.4
PySocks==1.7.1
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
rsa==4.9
selenium==4.25.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
tqdm==4.66.5
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
websocket-client==1.8.0
wsproto==1.2.0
